Validated the inputs in the authentication fragments, also moved all classes to the right packages for ease and clean work.
