+++ app/build.gradle
@@ -59,4 +59,10 @@ dependencies {
 
     def nav_version = "2.5.3"
     implementation "androidx.navigation:navigation-ui-ktx:$nav_version"
+
+    implementation "com.squareup.retrofit2:retrofit:2.9.0"
+    implementation "com.squareup.retrofit2:converter-gson:2.9.0"
+    implementation "com.squareup.okhttp3:okhttp:4.9.0"
+    implementation "com.squareup.okhttp3:logging-interceptor:4.9.0"
+
 }
