--- app/src/main/java/com/ayotola/fundall_app/LoginActivity.kt
@@ -1,32 +1,0 @@
-package com.ayotola.fundall_app
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.ayotola.fundall_app.databinding.ActivityLoginBinding
-import com.ayotola.fundall_app.databinding.ActivitySignupBinding
-
-class LoginActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityLoginBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityLoginBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.cancelBtn.setOnClickListener {
-            finish()
-        }
-
-        binding.loginButton.setOnClickListener {
-            binding.password.text.toString()
-            val intent = Intent(this, MainActivity::class.java)
-            startActivity(intent)
-        }
-
-        binding.createBtn.setOnClickListener {
-            val intent = Intent(this, SignupActivity::class.java)
-            startActivity(intent)
-        }
-    }
-}
