+++ app/src/main/java/com/ayotola/fundall_app/home/CardsFragment.kt
@@ -1,38 +1,29 @@
-package com.ayotola.fundall_app
+package com.ayotola.fundall_app.home
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.ayotola.fundall_app.adapter.CustomAdapter
 import com.ayotola.fundall_app.databinding.FragmentCardsBinding
-import com.ayotola.fundall_app.databinding.FragmentPickNewCardBinding
 
-
-class PickNewCardFragment : Fragment() {
-    private lateinit var binding: FragmentPickNewCardBinding
+class CardsFragment : Fragment() {
+    private lateinit var binding: FragmentCardsBinding
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        binding = FragmentPickNewCardBinding.inflate(inflater, container, false)
+    ): View {
+        binding = FragmentCardsBinding.inflate(inflater, container, false)
         return binding.root
     }
 
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        binding.backText.setOnClickListener {
-
-        }
-
-        binding.include.continueBtn.setOnClickListener {
-
-        }
-
 
         val recyclerView = binding.recyclerView
 
@@ -44,9 +35,5 @@ class PickNewCardFragment : Fragment() {
         val adapter = CustomAdapter(itemList)
         recyclerView.adapter = adapter
 
-        val customDialogFragment = DialogFragment()
-        fragmentManager?.let { customDialogFragment.show(it, "CustomDialogFragment") }
-
     }
-
 }
