+++ app/src/main/res/drawable/logo.xml
@@ -1,0 +1,17 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="88dp"
+    android:height="67dp"
+    android:viewportWidth="88"
+    android:viewportHeight="67">
+  <path
+      android:pathData="M33,66.165C24.248,66.165 15.854,62.68 9.665,56.476C3.477,50.271 0,41.857 0,33.083L0,0H22V33.083C22,36.007 23.159,38.812 25.222,40.88C27.285,42.948 30.083,44.11 33,44.11C35.917,44.11 38.715,42.948 40.778,40.88C42.841,38.812 44,36.007 44,33.083V0H66V33.083C66,41.857 62.523,50.271 56.334,56.476C50.146,62.68 41.752,66.165 33,66.165Z"
+      android:fillColor="#4DE897"/>
+  <path
+      android:pathData="M88,44.111H33V66.166H88V44.111Z"
+      android:fillColor="#016938"/>
+  <path
+      android:pathData="M33,66.165C24.248,66.165 15.854,62.68 9.665,56.476C3.477,50.271 0,41.857 0,33.083L0,0H22V33.083C22,36.007 23.159,38.812 25.222,40.88C27.285,42.948 30.083,44.11 33,44.11C35.917,44.11 38.715,42.948 40.778,40.88C42.841,38.812 44,36.007 44,33.083V0H66V33.083C66,41.857 62.523,50.271 56.334,56.476C50.146,62.68 41.752,66.165 33,66.165Z"
+      android:strokeAlpha="0.5"
+      android:fillColor="#4DE897"
+      android:fillAlpha="0.5"/>
+</vector>
