+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/background"
-    tools:context=".SplashScreenActivity">
+    tools:context=".authentication.SplashScreenActivity">
 
     <ImageView
         android:id="@+id/app_logo"
