+++ app/src/main/res/layout/fragment_dialog.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".DialogFragment">
+    tools:context=".dialog.DialogFragment">
 
     <ImageView
         android:id="@+id/dialog_exit_btn"
